requirements.txt
absl-py>=0.5.0
numpy>=1.13.3
tensorflow>=1.13.0
tensorflow-serving-api>=1.13.0
gin-config
pybullet
Pillow==5.3.0
gym>=0.10.9
tensorflow-probability>=0.6.0
